b/requirements.txt
# VerifiedCC Python Backend Dependencies

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
supabase==2.3.0

# Data validation
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0

# Guardian integration
requests==2.31.0
python-dateutil==2.8.2

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-mock==3.12.0